Rename tensor_comparison/core.py to compare.py
