+++ python/sglang/srt/debug_utils/comparator/tensor_comparison/__init__.py
@@ -1,4 +1,4 @@
-from sglang.srt.debug_utils.comparator.tensor_comparison.core import compare_tensors
+from sglang.srt.debug_utils.comparator.tensor_comparison.compare import compare_tensors
 from sglang.srt.debug_utils.comparator.tensor_comparison.printer import (
     print_comparison,
 )
